Use `indexPathForItemAtPoint:` for long press.
`handleLongPressGesture:` was using `indexPathForItemClosestToPoint:` method to
retrieve the index path of the pressed cell. This works well on layouts
that fill the entire collection view space but not on other layouts: in the
latter case, the closest cell will be selected whereas this is the one under the
finger we want to be selected.

It now uses `indexPathForItemAtPoint:` instead of
`indexPathForItemClosestToPoint:` to get the touched cell.